dts infinite loop bugfix

+++ include/pushplan/agents/robot.hpp
@@ -87,7 +87,7 @@ public:
 		const std::vector<Object*>& other_movables,
 		const comms::ObjectsPoses& rearranged,
 		comms::ObjectsPoses& result,
-		int& push_reward);
+		double& push_reward);
 	trajectory_msgs::JointTrajectory GetLastPlan()
 	{
 		m_planner->ProfilePath(m_rm.get(), m_traj);
